+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add packer-node component support
+
 ## 1.1.0 - 2026-09-23
 ### Added
 - Add update-examples capability.